+++ FrontEnd/pages/__init__.py
@@ -7,7 +7,6 @@ from typing import Callable
 
 from .customer_insights import render_customer_insight_tab
 from .dashboard import render_dashboard_tab
-from .live_stream import render_live_tab
 from .system_health import render_system_health_tab
 from .woocommerce import render_woocommerce_tab
 from FrontEnd.utils.config import PRIMARY_PAGE_CONFIG
@@ -23,7 +22,6 @@ class PrimaryPage:
 
 _PAGE_RENDERERS = {
     "business_intelligence": render_dashboard_tab,
-    "stream_monitor": render_live_tab,
     "customer_intelligence": render_customer_insight_tab,
     "commerce_hub": render_woocommerce_tab,
     "system_health": render_system_health_tab,
@@ -46,7 +44,6 @@ __all__ = [
     "PrimaryPage",
     "get_primary_pages",
     "render_dashboard_tab",
-    "render_live_tab",
     "render_customer_insight_tab",
     "render_woocommerce_tab",
     "render_system_health_tab",